fix: gracefully fail reports if elements or projects don't exist naymore
