+++ frontend/src/lib/components/report/Banner.svelte
@@ -1,0 +1,67 @@
+<div class="relative mt-10 w-full rounded bg-primary p-5">
+	<h4 class="text-xl text-white">Enjoyed this report? You can make one too!</h4>
+	<p class="mt-2 text-white">
+		Error analysis, chart authoring, shareable reports, and more with <a
+			class="font-bold transition hover:text-primary-dark"
+			href="https://zenoml.com">Zeno</a
+		>.
+	</p>
+	<div class="mt-4">
+		<button
+			class="rounded border border-white px-4 py-2 font-bold text-white transition hover:bg-primary-dark"
+			on:click={() => (window.location.href = 'https://zenoml.com')}
+		>
+			Learn more
+		</button>
+		<button
+			class="ml-4 rounded border border-white px-4 py-2 font-bold text-white transition hover:bg-primary-dark"
+			on:click={() => (window.location.href = 'https://zenoml.com/docs/intro')}
+		>
+			Get started
+		</button>
+	</div>
+	<div class="absolute right-0 top-0 hidden sm:inline">
+		<svg
+			width="300"
+			height="150"
+			viewBox="0 0 862 380"
+			fill="none"
+			xmlns="http://www.w3.org/2000/svg"
+		>
+			<path
+				d="M861.702 424.838L436.864 424.838L436.864 0.000147393L650.895 214.031L861.702 424.838Z"
+				fill="#9F55CD"
+			>
+				<animate attributeName="opacity" values=".6;1;.6" dur="5s" repeatCount="indefinite" />
+			</path>
+			<path
+				d="M417.842 426.106L205.423 213.687L417.842 1.26808L417.842 426.106Z"
+				fill="#9F55CD"
+				fill-opacity="0.8"
+			>
+				<animate attributeName="opacity" values=".8;.5;.8" dur="5s" repeatCount="indefinite" />
+			</path>
+			<path
+				d="M392.478 426.106L193.375 426.106L193.375 227.003L392.478 426.106Z"
+				fill="#9F55CD"
+				fill-opacity="0.6"
+			>
+				<animate attributeName="opacity" values=".4;.8;.4" dur="5s" repeatCount="indefinite" />
+			</path>
+			<path
+				d="M175.776 425.95L85.1012 335.276L176.255 244.122L175.776 425.95Z"
+				fill="#9F55CD"
+				fill-opacity="0.5"
+			>
+				<animate attributeName="opacity" values=".6;.2;.6" dur="5s" repeatCount="indefinite" />
+			</path>
+			<path
+				d="M8.92166e-05 417.575L68.7054 348.87L143.13 423.294L8.92166e-05 417.575Z"
+				fill="#9F55CD"
+				fill-opacity="0.5"
+			>
+				<animate attributeName="opacity" values=".2;.4;.2" dur="5s" repeatCount="indefinite" />
+			</path>
+		</svg>
+	</div>
+</div>
