+++ frontend/src/lib/components/report/elements/ChartElement.svelte
@@ -10,23 +10,31 @@
 </script>
 
 <div class="w-full">
-	<h3 class="text-lg font-semibold">{chart.name}</h3>
-	{#await zenoClient.getChartConfig(chart.projectUuid, chart.id) then chartConfig}
-		{#await zenoClient.getChartData(chart.projectUuid, chart.id) then data}
-			<div class="text-center">
-				<svelte:component
-					this={chartMap[chart.type]}
-					{chart}
-					{chartConfig}
-					{width}
-					data={JSON.parse(data)}
-					height={chart.type == ChartType.RADAR ? 600 : 400}
-				/>
-			</div>
+	{#if chart === undefined}
+		<p class="ml-4 mt-4 font-semibold text-error">Chart does not exist anymore.</p>
+	{:else}
+		<h3 class="text-lg font-semibold">{chart.name}</h3>
+		{#await zenoClient.getChartConfig(chart.projectUuid, chart.id) then chartConfig}
+			{#await zenoClient.getChartData(chart.projectUuid, chart.id) then data}
+				<div class="text-center">
+					<svelte:component
+						this={chartMap[chart.type]}
+						{chart}
+						{chartConfig}
+						{width}
+						data={JSON.parse(data)}
+						height={chart.type == ChartType.RADAR ? 600 : 400}
+					/>
+				</div>
+			{:catch error}
+				<p class="ml-4 mt-4 font-semibold text-error">
+					Chart data could not be loaded: {error.message}
+				</p>
+			{/await}
 		{:catch error}
 			<p class="ml-4 mt-4 font-semibold text-error">
-				Chart data could not be loaded: {error.message}
+				Chart config could not be loaded: {error.message}
 			</p>
 		{/await}
-	{/await}
+	{/if}
 </div>
